backend/app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
passlib[bcrypt]
PyJWT
python-dotenv
pytest
httpx
pytest-asyncio
